requirements.txt
Django==2.2.24
pinax-templates==2.0.6
pinax-documents==2.0.0
django-crispy-forms
pinax-messages==3.0.0
pinax-notifications==5.0.3
django-user-accounts==3.0.4
Django-appconf==1.0.4
metron==1.3.7
pinax-eventlog==4.0.1
django-localflavor==2.2
phonenumbers
django-phonenumber-field
sqlobject
ipython
xlwt
mysqlclient
django-mysql
django-fixture-magic
django-wkhtmltopdf
unidecode
#ipython==5
#gunicorn==19.6.0
django-anymail
jsonfield==2.1.1
python-dateutil
pycrypto
bootstrap3-datetime
django-cookie-consent
django-excel-response
